+++ sdk/core/core/inc/_internal/az_mock_curl.h
@@ -5,7 +5,7 @@
 #define _az_MOCK_CURL_H
 
 #include <_internal/az_http_request.h>
-#include <_internal/az_http_response.h>
+#include <az_http_response.h>
 #include <az_result.h>
 
 #include <stdbool.h>
