+++ sdk/core/core/inc/az_http_response.h
