+++ sdk/core/core/inc/az_span_reader.h
