+++ sdk/core/core/src/az_http_pipeline.c
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-#include <_internal/az_http_pipeline.h>
-#include <_internal/az_http_policy.h>
+#include <az_http_pipeline.h>
+#include <az_http_policy.h>
 
 #include <_az_cfg.h>
