+++ sdk/keyvault/keyvault/inc/az_keyvault.h
@@ -4,8 +4,8 @@
 #ifndef _az_KEYVAULT_H
 #define _az_KEYVAULT_H
 
-#include <_internal/az_http_pipeline.h>
-#include <_internal/az_http_response.h>
+#include <az_http_pipeline.h>
+#include <az_http_response.h>
 #include <az_contract.h>
 #include <az_identity_access_token.h>
 #include <az_identity_access_token_context.h>
